reverted back to /details/:id endpoint + added morgan
