+++ server/app.js
@@ -1,15 +1,19 @@
 const express = require('express');
 const bodyParser = require('body-parser');
 const Room = require('../database/models/room.js');
+const morgan = require('morgan');
+
 
 let app = express();
+
+app.use(morgan('dev'));
 
 app.use(bodyParser.urlencoded({extended: true}));
 app.use(bodyParser.json());
 
 app.use(express.static(__dirname + '/../client/dist'));
 
-app.get('/:id/details', (req, res) => {
+app.get('/details/:id', (req, res) => {
   Room.findByID(req.params.id, (err, roomInfo) => {
     if (err) {
       res.status(404).json({error: `ID ${req.params.id} does not exist in database`});
