+++ server/app.test.js
@@ -30,19 +30,19 @@ describe('server', () => {
     });
   });
 
-  describe('requests to /rooms/:id', () => {
+  describe('requests to /details/:id', () => {
     test('it should return JSON with a \'data\' key for route \'1/details\'', (done) => {
       Room.findByID.mockImplementation((id, cb) => {
         cb(null, [{id: 'this is a test'}]);
       });
-      request.get('/1/details').then((response) => {
+      request.get('/details/1').then((response) => {
         expect(response.body).toHaveProperty('data');
         done();
       });
     });
 
     test('it should only return a single item inside the data key', (done) => {
-      request.get('/1/details').then((response) => {
+      request.get('/details/1').then((response) => {
         expect(response.body.data.length).toBe(1);
         done();
       });
@@ -52,14 +52,14 @@ describe('server', () => {
       Room.findByID.mockImplementation((id, cb) => {
         cb('error', null);
       });
-      request.get('/1/details').then((response) => {
+      request.get('/details/1').then((response) => {
         expect(response.statusCode).toBe(404);
         done();
       });
     });
 
     test('response message for the invalid message should contain the id', (done) => {
-      request.get('/1/details').then((response) => {
+      request.get('/details/1').then((response) => {
         expect(response.body.error).toBe('ID 1 does not exist in database');
         done();
       });
